### 2020-01-17 #2024 @whedon
[PRE REVIEW]: SpeciesEvolver: A Landlab component to evolve life in simulated landscapes
**Submitting author:** @nathanlyons (<a href="http://orcid.org/0000-0001-6965-3374">Nathan Lyons</a>)
**Repository:** <a href="https://github.com/landlab/landlab" target ="_blank">https://github.com/landlab/landlab</a>
**Version:** v2.0.0b3
