### 2020-01-17 #2024 @kyleniemeyer
Hi @nathanlyons, could you clarify the focus of your submission (SpeciesEvolver) vs the main package of the repository you listed, Landlab? At the repo, I see an article for "GroundwaterDupuitPercolator".
