### 2020-01-17 #2024 @nathanlyons
Hi @kyleniemeyer, the article "GroundwaterDupuitPercolator" on the master branch is also in JOSS review. This submission is on the branch, `nathanlyons_species_evolver_joss_review` where the paper materials live in /joss/species_evolver.
