### 2020-01-20 #2024 @nathanlyons
Thanks for the notice, @kyleniemeyer.

I believe I understand the reasoning of tagging kbarnhark and I have a sense of the role of the handling editor. May you please describe the anticipated role of an additional member of the editorial team? I ask just so I can understand the process.
