### 2020-01-22 #2024 @kyleniemeyer
@nathanlyons it probably isn't necessary, but mostly just to help me understand the relationship/role of Landlab and its components
