### 2020-02-03 #2024 @kyleniemeyer
Hi @ebgoldstein, would you be willing to review this submission for JOSS?
