### 2020-02-04 #2066 @nathanlyons
@kyleniemeyer @ebgoldstein @fclubb Thank you in advance for your time.

**The Landlab branch under this review is `nathanlyons_species_evolver_joss_review`.** This branch will be merged with the Landlab development branch, `master` then `release` following this review. The separate joss review branch is necessary because multiple Landlab modules are under JOSS review.
