### 2020-02-04 #2066 @kyleniemeyer
Thanks @nathanlyons, I forgot about the paper being on a branch—I've deleted the incorrect PDF.
