### 2020-02-08 landlab/landlab#1147 @fclubb
SpeciesEvolver JOSS review
Hi @nathanlyons,
Just a couple of suggestions for the JOSS review (I think I am supposed to do this in an issue here, apologies if that's not the case). I think this is a really great component and I look forward to using it for both research and teaching in the future!

In the software paper:
- Description of zone connectivity relationships: just to clarify - does many-to-many require the same number of zones in each timestep? In Fig 1b two patches in the first time step remain as two patches in the second time step. Could two patches become three, etc.?

Install:
- I installed from source, but I presume that you once this component is merged with the `master` branch then `SpeciesEvolver` will be available through the pre-packaged binary?
- The install went very smoothly and I worked through the jupyter notebook with no problems. This is a very cool model!

Documentation:
- The JOSS review checklist has "statement of need" in the documentation section. While you have a clear statement of need in the JOSS paper, I couldn't find this on the SpeciesEvolver documentation (on ReadTheDocs). You could update the docs with a couple of sentences highlighting the purpose of the component and the target audience.
- The documentation for each class and function looks very thorough and I had no problems understanding it.
