### 2020-02-08 #2066 @fclubb
Hi @nathanlyons, I have updated my checklist and made a couple of very minor suggestions in a Landlab issue: https://github.com/landlab/landlab/issues/1147
