### 2020-02-08 landlab/landlab#1147 @nathanlyons
Hi @fclubb,

Thanks for the comments! Two comments required changes to the paper and code. In these cases, a link to the commit addressing the comment is provided. The commits were pushed to `nathanlyons_species_evolver_joss_review`.

> I look forward to using it for both research and teaching in the future!

#FavoriteThingToReadInAReview

> Description of zone connectivity relationships: just to clarify - does many-to-many require the same number of zones in each timestep? In Fig 1b two patches in the first time step remain as two patches in the second time step. Could two patches become three, etc.?

Good question that should be addressed in the paper. In the `many-to-many` bullet of the paper, I added:

`Zone counts in both the prior and current steps must be greater than one for the connectivity to be assigned this relationship. Zone counts in these steps do not need to be the same for the connectivity to be assigned this relationship.`

In the many-to-many of Fig 1b, I added a variation on the current time step showing 3 zones where the prior step had 2 zones. To summarize, many-to-many is any combination of prior and current counts when the counts are greater than 1.

addressed in [this commit](https://github.com/landlab/landlab/commit/d2e6395c96ad19a658d5f28bac9ccf86c19598c4)

> I installed from source, but I presume that you once this component is merged with the master branch then SpeciesEvolver will be available through the pre-packaged binary?

Coincidently, Landlab 2 beta is the most recent pre-packaged binary, and this includes SpeciesEvolver. Meaning, SpeciesEvolver is already included in Landlab (2 beta), through conda and pip, as described in the standard [installation instructions](https://landlab.readthedocs.io/en/release/install/index.html). SpeciesEvolver revisions made during this JOSS review will rapidly make their way into the binaries following review completion.

> The install went very smoothly and I worked through the jupyter notebook with no problems. This is a very cool model!

So glad to read the installation went smoothly. The Landlab team has worked to make it smooth.

Also good to read you enjoyed the Jupyter notebook. The intent was to demonstrate SpeciesEvolver in a straightforward and interesting way.

> The JOSS review checklist has "statement of need" in the documentation section. While you have a clear statement of need in the JOSS paper, I couldn't find this on the SpeciesEvolver documentation (on ReadTheDocs). You could update the docs with a couple of sentences highlighting the purpose of the component and the target audience.

Very much agree that a sort of module abstract is needed. I included this in two places:

1. At the top of the SpeciesEvolver ReadTheDocs page. Because the Landlab ReadTheDocs is not built for the review branch, this change will not be visible on ReadTheDocs until the branch is merged with master, release. Until then, the change can be seen at the top of the [component file](https://github.com/landlab/landlab/blob/nathanlyons_species_evolver_joss_review/landlab/components/species_evolution/species_evolver.py).
2. The readme displayed in the [component directory](https://github.com/landlab/landlab/tree/nathanlyons_species_evolver_joss_review/landlab/components/species_evolution).

addressed in [this commit](https://github.com/landlab/landlab/commit/0f54b720a2d12761527d7c4a46810e7fedba6fac)

> The documentation for each class and function looks very thorough and I had no problems understanding it.

Ecstatic to read the documentation effort delivered!

(I will keep this issue open while SpeciesEvolver is under JOSS review.)
