### 2020-02-08 #2066 @nathanlyons
@fclubb I have responded to your comments at https://github.com/landlab/landlab/issues/1147. Code changes were pushed to the [branch under review](https://github.com/landlab/landlab/tree/nathanlyons_species_evolver_joss_review). I will ask the robot to create an updated version of the paper reflecting your comments.
