### 2020-02-10 a92375c
remove unused modules from landlab.grid.structured_quad
