+++ tests/grid/structured_quad/test_links.py
@@ -4,7 +4,6 @@ from numpy.testing import assert_array_equal
 
 from landlab.grid.nodestatus import NodeStatus
 from landlab.grid.structured_quad.links import active_link_ids
-from landlab.grid.structured_quad.nodes import status_with_perimeter_as_boundary
 from landlab.testing.tools import assert_array_is_int
 
 
@@ -15,14 +14,6 @@ def test_active_links_ids():
 
     link_ids = active_link_ids((4, 5), status)
     assert_array_equal(link_ids, [11, 15, 16, 20])
-    # assert_array_equal(link_ids, [7, 8, 21, 25])
-    assert_array_is_int(link_ids)
-
-
-def test_active_links_with_edge_boundaries():
-    status = status_with_perimeter_as_boundary((3, 4))
-    link_ids = active_link_ids((3, 4), status)
-    assert_array_equal(link_ids, [8])
     assert_array_is_int(link_ids)
 
 
