### 2020-02-11 #2066 @ebgoldstein
👋  @nathanlyons

I have now worked through the Species Evolver [tutorial notebook](https://github.com/landlab/landlab/tree/master/notebooks/tutorials/species_evolution) (via a Google Colab).  The install and model worked without a hiccup — i think this component extends the capabilities of the model in exciting ways, and will allow many interesting questions to be addressed using Landlab.

The paper was clear, and I understood both the SON and the summary of the component.

@kyleniemeyer, at this time I have made no issues in the Landlab repository, and i have no suggested edits to the ms.

very cool work @nathanlyons et al.
