### 2020-02-11 83f34d7
first steps and some notes
