### 2020-02-12 5e30d92
Update with JOSS pub details
