### 2020-02-12 b84a821
made paper corrections
