+++ joss/species_evolver/paper.bib
@@ -96,7 +96,7 @@
     isbn={0716710927},
     publisher={WH Freeman},
     title = {Macroevolution: pattern and process},
-    year = {1993}
+    year = {1979}
 }
 
 @article{Taylor:1993,
