### 2020-02-12 #2066 @nathanlyons
@ebgoldstein Thank you also for the review. Good to know it works in Google Colab. I suppose no reason it should not.
