### 2020-02-12 #2066 @nathanlyons
@kyleniemeyer FYI for your review, I noticed minor changes needed in the paper, namely an incorrect year on a reference, acknowledging all reviewers, and rewording "taxa evolution" to "biological evolution" because taxa was not defined yet in the Introduction. Here is the [commit](https://github.com/landlab/landlab/commit/b84a8214ed4a99ed2c9cca2c048052ac5eddb153) with these changes, along with an additional word change in [here](https://github.com/landlab/landlab/commit/c2e815be02b7bf0a762ad5a32570ece4abfdf13b). I will generate an updated pdf.

In the meantime and given this repo is a branch, I foresee that this is the next course of action for me, at least given this recent precedent: https://github.com/openjournals/joss-reviews/issues/1935
- I merge my Landlab review branch with Landlab master (first I have requested groundwater's now published paper to no longer be named 'paper.md' so as to not conflict with mine.
- A new landlab version is released, maybe this will be v2.0.0b6, I will let you know.
- I set up an archive at zenodo and provide this doi to you.

@kyleniemeyer Please let me know if this is off.
