### 2020-02-12 #2066 @kyleniemeyer
@nathanlyons yes, that seems a good course of action. However, before you merge your branch into the Landlab master, please wait until i can review your article—that way any changes are contained in the branch
