### 2020-02-12 #2066 @kyleniemeyer
@nathanlyons just a minor comment: please consistently put commas after instances of "i.e." and "e.g." (for example, `(i.e. allopatry)` -> `(i.e., allopatry)`)
