### 2020-02-12 607f5c1
add unit agnostic to component base class
