+++ landlab/core/model_component.py
@@ -8,6 +8,7 @@ Base component class methods
 
     ~landlab.core.model_component.Component.name
     ~landlab.core.model_component.Component.from_path
+    ~landlab.core.model_component.Component.unit_agnostic
     ~landlab.core.model_component.Component.units
     ~landlab.core.model_component.Component.definitions
     ~landlab.core.model_component.Component.input_var_names
@@ -40,6 +41,7 @@ name: {name}
 description:
 {desc}
 units: {units}
+unit agnostic: {unit_agnostic}
 at: {loc}
 intent: {intent}
 """
@@ -56,7 +58,8 @@ class Component:
     _info = {}
     _name = None
     _cite_as = ""
-
+    _unit_agnostic = None
+    
     def __new__(cls, *args, **kwds):
         registry.add(cls)
         return object.__new__(cls)
@@ -252,6 +255,25 @@ class Component:
 
     @classproperty
     @classmethod
+    def unit_agnostic(cls):
+        """Whether the component is unit agnostic.
+
+        If True, then the component is unit agnostic. Under this condition a
+        user must still provide consistent units across all input arguments,
+        keyword arguments, and fields. However, when ``unit_agnostic`` is True
+        the units specified can be interpreted as dimensions.
+
+        When False, then the component requires inputs in the specified units.
+
+        Returns
+        -------
+        bool
+
+        """
+        return cls._unit_agnostic
+
+    @classproperty
+    @classmethod
     def units(cls):
         """Get the units for all field values.
 
@@ -329,7 +351,12 @@ class Component:
         intent = cls._info[name]["intent"]
 
         help = _VAR_HELP_MESSAGE.format(
-            name=name, desc=desc, units=units, loc=loc, intent=intent
+            name=name,
+            desc=desc,
+            units=units,
+            loc=loc,
+            intent=intent,
+            unit_agnostic=cls._unit_agnostic,
         )
 
         print(help.strip())
