+++ landlab/components/stream_power/stream_power_smooth_threshold.py
@@ -100,6 +100,8 @@ class StreamPowerSmoothThresholdEroder(FastscapeEroder):
 
     _name = "StreamPowerSmoothThresholdEroder"
 
+    _unit_agnostic = True
+    
     _cite_as = """
     @article{barnhart2019terrain,
       author = {Barnhart, Katherine R and Glade, Rachel C and Shobe, Charles M and Tucker, Gregory E},
@@ -331,7 +333,9 @@ class StreamPowerSmoothThresholdEroder(FastscapeEroder):
                 * self._A_to_the_m[defined_flow_receivers]
             ) / (self._thresholds[defined_flow_receivers] * flow_link_lengths)
 
-            self._delta[defined_flow_receivers][self._thresholds[defined_flow_receivers] == 0.0] = 0.0
+            self._delta[defined_flow_receivers][
+                self._thresholds[defined_flow_receivers] == 0.0
+            ] = 0.0
         else:
             self._gamma[defined_flow_receivers] = dt * self._thresholds
             if self._thresholds == 0:
