+++ landlab/components/pet/potential_evapotranspiration_field.py
@@ -90,6 +90,8 @@ class PotentialEvapotranspiration(Component):
 
     _name = "PotentialEvapotranspiration"
 
+    _unit_agnostic = False
+    
     _info = {
         "radiation__incoming_shortwave_flux": {
             "dtype": float,
