+++ landlab/components/depth_dependent_diffusion/hillslope_depth_dependent_linear_flux.py
@@ -119,6 +119,8 @@ class DepthDependentDiffuser(Component):
 
     _name = "DepthDependentDiffuser"
 
+    _unit_agnostic = True
+    
     _cite_as = """
     @article{barnhart2019terrain,
       author = {Barnhart, Katherine R and Glade, Rachel C and Shobe, Charles M and Tucker, Gregory E},
