+++ landlab/components/fire_generator/generate_fire.py
@@ -88,6 +88,8 @@ class FireGenerator(Component):
 
     _name = "FireGenerator"
 
+    _unit_agnostic = True
+    
     _info = {}
 
     def __init__(
