+++ landlab/components/weathering/exponential_weathering.py
@@ -62,6 +62,8 @@ class ExponentialWeatherer(Component):
 
     _name = "ExponentialWeatherer"
 
+    _unit_agnostic = True
+    
     _cite_as = """
     @article{barnhart2019terrain,
       author = {Barnhart, Katherine R and Glade, Rachel C and Shobe, Charles M and Tucker, Gregory E},
