+++ tests/components/test_components.py
@@ -23,8 +23,10 @@ def test_component_metadata(Comp):
         "SoilMoisture",
         "Vegetation",
     ):
-        print(Comp.name)
+        #print(Comp.name)
         grid = RasterModelGrid((10, 10))
+
+        assert Comp._unit_agnostic in (True, False)
 
         # verify that we can create it
         for name in Comp._info.keys():
