+++ docs/source/user_guide/index.rst
@@ -51,6 +51,15 @@ Model with Landlab and Components
     components
     build_a_model
 
+Landlab and Units
+-----------------
+
+.. toctree::
+    :maxdepth: 2
+
+    units
+
+
 Landlab Tutorial Library
 ------------------------
 
