+++ landlab/components/species_evolution/species_evolver.py
@@ -19,6 +19,7 @@ import numpy as np
 from pandas import DataFrame, isnull
 
 from landlab import Component
+
 from .record import Record
 
 
@@ -228,7 +229,7 @@ class SpeciesEvolver(Component):
 
     """
 
-    _name = 'SpeciesEvolver'
+    _name = "SpeciesEvolver"
 
     _info = {
         "taxa__richness": {
@@ -238,7 +239,7 @@ class SpeciesEvolver(Component):
             "units": "-",
             "mapping": "node",
             "doc": "The number of taxa at each node",
-        },
+        }
     }
 
     _cite_as = """@article{lyons2020species,
@@ -263,21 +264,21 @@ class SpeciesEvolver(Component):
         # Create data structures.
 
         self._record = Record(initial_time)
-        self._record.set_value('taxa', 0)
+        self._record.set_value("taxa", 0)
 
-        self._taxa = OrderedDict([
-            ('uid', []),
-            ('appeared', []),
-            ('latest_time', []),
-            ('extant', []),
-            ('object', [])]
+        self._taxa = OrderedDict(
+            [
+                ("uid", []),
+                ("appeared", []),
+                ("latest_time", []),
+                ("extant", []),
+                ("object", []),
+            ]
         )
 
         # Create a taxa richness field.
 
-        _ = grid.add_zeros(
-            'taxa__richness', at='node', dtype=int, clobber=True
-        )
+        _ = grid.add_zeros("taxa__richness", at="node", dtype=int, clobber=True)
 
     @property
     def record_data_frame(self):
@@ -312,10 +313,10 @@ class SpeciesEvolver(Component):
         The DataFrame is created from a data structure within the component.
         """
         cols = list(self._taxa.keys())
-        cols.remove('uid')
-        cols.remove('object')
-        df = DataFrame(self._taxa, columns=cols, index=self._taxa['uid'])
-        df.index.name = 'uid'
+        cols.remove("uid")
+        cols.remove("object")
+        df = DataFrame(self._taxa, columns=cols, index=self._taxa["uid"])
+        df.index.name = "uid"
 
         return df
 
@@ -336,7 +337,7 @@ class SpeciesEvolver(Component):
 
         # Create a dictionary of the taxa to update, `taxa_evolving`.
 
-        extant_taxa = np.array(self._taxa['object'])[self._taxa['extant']]
+        extant_taxa = np.array(self._taxa["object"])[self._taxa["extant"]]
         taxa_evolving = OrderedDict.fromkeys(extant_taxa, True)
 
         # Run stages of taxa evolution.
@@ -437,7 +438,7 @@ class SpeciesEvolver(Component):
         """
         time = self._record.latest_time
 
-        t_recorded = self._taxa['object']
+        t_recorded = self._taxa["object"]
 
         # Update previously introduced taxa.
 
@@ -446,9 +447,9 @@ class SpeciesEvolver(Component):
         for taxon in t_introduced:
             # Update taxon data.
 
-            idx = self._taxa['object'].index(taxon)
-            self._taxa['latest_time'][idx] = time
-            self._taxa['extant'][idx] = taxon.extant
+            idx = self._taxa["object"].index(taxon)
+            self._taxa["latest_time"][idx] = time
+            self._taxa["extant"][idx] = taxon.extant
 
         # Set the data of new taxa.
 
@@ -457,28 +458,27 @@ class SpeciesEvolver(Component):
         for taxon in t_new:
             # Set identifiers.
 
-            if self._taxa['uid']:
-                taxon._uid = max(self._taxa['uid']) + 1
+            if self._taxa["uid"]:
[…]
