### 2020-02-13 #2066 @nathanlyons
@kyleniemeyer All `i.e.` and `e.g.` instances are now followed by commas. I will now work on the aforementioned workflow.
