+++ landlab/components/species_evolution/species_evolver.py
@@ -198,7 +198,7 @@ class SpeciesEvolver(Component):
     behavior can be achieved by subclassing ``ZoneTaxon`` or ``Taxon``.
     """
     _name = 'SpeciesEvolver'
-    
+
     _unit_agnostic = True
 
     _info = {
