+++ landlab/components/taylor_nonlinear_hillslope_flux/taylor_nonlinear_hillslope_flux.py
@@ -142,7 +142,7 @@ class TaylorNonLinearDiffuser(Component):
     _name = "TaylorNonLinearDiffuser"
 
     _unit_agnostic = True
-    
+
     _cite_as = """
     @article{barnhart2019terrain,
       author = {Barnhart, Katherine R and Glade, Rachel C and Shobe, Charles M and Tucker, Gregory E},
