+++ landlab/components/erosion_deposition/generalized_erosion_deposition.py
@@ -15,7 +15,7 @@ class _GeneralizedErosionDeposition(Component):
     _name = "_GeneralizedErosionDeposition"
 
     _unit_agnostic = True
-    
+
     _info = {
         "flow__link_to_receiver_node": {
             "dtype": int,
