+++ landlab/components/diffusion/diffusion.py
@@ -107,7 +107,7 @@ class LinearDiffuser(Component):
     _name = "LinearDiffuser"
 
     _unit_agnostic = True
-    
+
     _info = {
         "hillslope_sediment__unit_volume_flux": {
             "dtype": float,
