+++ landlab/components/hack_calculator/hack_calculator.py
@@ -172,7 +172,7 @@ class HackCalculator(Component):
     _name = "HackCalculator"
 
     _unit_agnostic = True
-    
+
     _info = {
         "distance_to_divide": {
             "dtype": float,
