### 2020-02-17 4db7f0d
merge for v2.0.0b6 (#1153)
* docs: fix typos (#1145)

* Update with JOSS pub details

* move JOSS paper to published

* add cite_as attribute

* species evolver joss review branch (#1152)

* initial commit

* add joss figure

* corrected reference

* corrected indention

* added space before list

* added acknowledgements

* removed figure section

* test paper

* corrected author name character

* remove figure

* remove isbn

* reduced words

* fixed affiliation indent

* removed list

* removed extra space

* changed figure format

* changed figure format to eps

* changed figure format to 
[…]
