+++ tests/components/species_evolution/test_record.py
@@ -15,36 +15,34 @@ def test_attributes():
 
     record.advance_time(10)
 
-    record.set_value('vara', 1)
-    record.set_value('varb', 2)
+    record.set_value("vara", 1)
+    record.set_value("varb", 2)
 
     np.testing.assert_equal(record.times, [0, 10])
     np.testing.assert_equal(record.prior_time, 0)
     np.testing.assert_equal(record.earliest_time, 0)
     np.testing.assert_equal(record.latest_time, 10)
-    np.testing.assert_equal(record.variables, ['vara', 'varb'])
+    np.testing.assert_equal(record.variables, ["vara", "varb"])
     np.testing.assert_equal(record.count_of_time_steps, 2)
     np.testing.assert_equal(len(record), 2)
 
-    df = pd.DataFrame(
-        {'time': [0, 10], 'vara': [np.nan, 1], 'varb': [np.nan, 2]}
-    )
+    df = pd.DataFrame({"time": [0, 10], "vara": [np.nan, 1], "varb": [np.nan, 2]})
     pd.testing.assert_frame_equal(record.data_frame, df, check_like=True)
 
 
 def test_get_value():
     record = Record()
-    record.set_value('vara', 1)
+    record.set_value("vara", 1)
     record.advance_time(10)
-    record.set_value('vara', 2)
+    record.set_value("vara", 2)
 
-    value = record.get_value('vara')
+    value = record.get_value("vara")
     np.testing.assert_equal(value, 2)
 
-    value = record.get_value('vara', time=0)
+    value = record.get_value("vara", time=0)
     np.testing.assert_equal(value, 1)
 
-    value = record.get_value('varb')
+    value = record.get_value("varb")
     np.testing.assert_equal(value, np.nan)
 
 
@@ -52,48 +50,46 @@ def test_set_value():
     record = Record()
     record.advance_time(10)
 
-    record.set_value('vara', 1)
-    d = OrderedDict([('time', [0, 10]), ('vara', [np.nan, 1])])
+    record.set_value("vara", 1)
+    d = OrderedDict([("time", [0, 10]), ("vara", [np.nan, 1])])
     np.testing.assert_equal(record._dict, d)
 
-    record.set_value('vara', 3, time=0)
-    d = OrderedDict([('time', [0, 10]), ('vara', [3, 1])])
+    record.set_value("vara", 3, time=0)
+    d = OrderedDict([("time", [0, 10]), ("vara", [3, 1])])
     np.testing.assert_equal(record._dict, d)
 
 
 def test_increment_value():
     record = Record()
-    record.set_value('vara', 1)
-    record.increment_value('vara', 100)
+    record.set_value("vara", 1)
+    record.increment_value("vara", 100)
 
-    d = OrderedDict([('time', [0]), ('vara', [101])])
+    d = OrderedDict([("time", [0]), ("vara", [101])])
     np.testing.assert_equal(record._dict, d)
 
     record.advance_time(10)
-    record.increment_value('vara', 100)
+    record.increment_value("vara", 100)
 
-    d = OrderedDict([('time', [0, 10]), ('vara', [101, 100])])
+    d = OrderedDict([("time", [0, 10]), ("vara", [101, 100])])
     np.testing.assert_equal(record._dict, d)
 
-    record.increment_value('varb', 200, time=0)
+    record.increment_value("varb", 200, time=0)
 
-    d = OrderedDict(
-        [('time', [0, 10]), ('vara', [101, 100]), ('varb', [200, np.nan])]
-    )
+    d = OrderedDict([("time", [0, 10]), ("vara", [101, 100]), ("varb", [200, np.nan])])
     np.testing.assert_equal(record._dict, d)
 
 
[…]
