+++ landlab/components/species_evolution/species_evolver.py
@@ -2,6 +2,14 @@
 # -*- coding: utf-8 -*-
 """Evolve life in a landscape.
 
+Life evolves alongside landscapes by biotic and abiotic processes under complex
+dynamics at Earth's surface. Researchers who wish to explore these dynamics can
+use this component as a tool for them to build landscape-life evolution models.
+Landlab components, including SpeciesEvolver are designed to work with a shared
+model grid. Researchers can build novel models using plug-and-play surface
+process components to evolve the grid's landscape alongside the life tracked by
+SpeciesEvolver. The simulated life evolves following customizable processes.
+
 Component written by Nathan Lyons beginning August 2017.
 """
 from collections import OrderedDict
@@ -10,6 +18,7 @@ import numpy as np
 from pandas import DataFrame, isnull
 
 from landlab import Component
+
 from .record import Record
 
 
@@ -196,8 +205,28 @@ class SpeciesEvolver(Component):
     ``ZoneTaxon`` as well as the parameters used in this example (default
     values were used because optional parameters were not set). Different
     behavior can be achieved by subclassing ``ZoneTaxon`` or ``Taxon``.
+
+    References
+    ----------
+    **Required Software Citation(s) Specific to this Component**
+
+    Lyons, N.J., Albert, J.S., Gasparini, N.M. (in review). SpeciesEvolver: A
+    Landlab component to evolve life in simulated landscapes. Journal of Open
+    Source Software.
+
+    **Additional References**
+
+    Albert, J.S., Schoolmaster Jr, D.R., Tagliacollo, V., Duke-Sylvester, S.M.
+    (2016). Barrier displacement on a neutral landscape: Toward a theory of
+    continental biogeography. Systematic Biology 66(2), 167–182.
+
+    Lyons, N.J., Val, P., Albert, J.S., Willenbring, J.K., Gasparini, N.M., in
+    review. Topographic controls on divide migration, stream capture, and
+    diversification in riverine life. Earth Surface Dynamics.
+
     """
-    _name = 'SpeciesEvolver'
+
+    _name = "SpeciesEvolver"
 
     _info = {
         "taxa__richness": {
@@ -207,8 +236,14 @@ class SpeciesEvolver(Component):
             "units": "-",
             "mapping": "node",
             "doc": "The number of taxa at each node",
-        },
+        }
     }
+
+    _cite_as = """@article{lyons2020species,
+        author = {Lyons, N.J. and Albert, J.S. and Gasparini, N.M.},
+        title = {SpeciesEvolver: A Landlab component to evolve life in simulated landscapes},
+        journal = {Journal of Open Source Software},
+        }"""
 
     def __init__(self, grid, initial_time=0):
         """Instantiate SpeciesEvolver.
@@ -226,21 +261,21 @@ class SpeciesEvolver(Component):
         # Create data structures.
 
         self._record = Record(initial_time)
-        self._record.set_value('taxa', 0)
+        self._record.set_value("taxa", 0)
 
-        self._taxa = OrderedDict([
-            ('uid', []),
-            ('appeared', []),
-            ('latest_time', []),
-            ('extant', []),
-            ('object', [])]
+        self._taxa = OrderedDict(
+            [
+                ("uid", []),
+                ("appeared", []),
+                ("latest_time", []),
+                ("extant", []),
+                ("object", []),
+            ]
         )
 
         # Create a taxa richness field.
 
-        _ = grid.add_zeros(
-            'taxa__richness', at='node', dtype=int, clobber=True
-        )
[…]
