### 2020-02-18 #2066 @nathanlyons
@kyleniemeyer

The software under review is no longer on a branch, and is now on `master` of the Landlab repo.

Please update the version of the software under review. It should now be `v2.0.0b6`.

The Zenodo archive DOI is `10.5281/zenodo.3672119`.
